Move GlossaryApi into its own file
