+++ lib/deepl.rb
@@ -37,6 +37,7 @@ require 'deepl/resources/usage'
 # -- Other wrappers
 require 'deepl/api'
 require 'deepl/configuration'
+require 'deepl/glossary_api'
 
 # -- Gem interface
 module DeepL
@@ -65,38 +66,6 @@ module DeepL
   def usage(options = {})
     configure if @configuration.nil?
     Requests::Usage.new(api, options).request
-  end
-
-  class GlossaryApi
-    def initialize(api, options = {})
-      @api = api
-      @options = options
-    end
-
-    def create(name, source_lang, target_lang, entries, entries_format = 'tsv', options = {})
-      DeepL::Requests::Glossary::Create.new(@api, name, source_lang, target_lang, entries,
-                                            entries_format, options).request
-    end
-
-    def destroy(glossary_id, options = {})
-      DeepL::Requests::Glossary::Destroy.new(@api, glossary_id, options).request
-    end
-
-    def entries(glossary_id, options = {})
-      DeepL::Requests::Glossary::Entries.new(@api, glossary_id, options).request
-    end
-
-    def find(glossary_id, options = {})
-      DeepL::Requests::Glossary::Find.new(@api, glossary_id, options).request
-    end
-
-    def language_pairs(options = {})
-      DeepL::Requests::Glossary::LanguagePairs.new(@api, options).request
-    end
-
-    def list(options = {})
-      DeepL::Requests::Glossary::List.new(@api, options).request
-    end
   end
 
   # -- Configuration
